[2.console] 1) draw pix, rectangel, character; 2) a console with a simple kprintf

+++ src/types.h
@@ -6,15 +6,14 @@
 #ifndef _TYPES_H_
 #define _TYPES_H_
 
-typedef unsigned char   uint8;
-typedef unsigned short  uint16;
-typedef unsigned int    uint32;
-typedef unsigned long   uint64;
-
-typedef char   int8;
-typedef short  int16;
-typedef int    int32;
-typedef long   int64;
+typedef unsigned char       uint8;
+typedef unsigned short      uint16;
+typedef unsigned int        uint32;
+typedef unsigned long long  uint64;
+typedef char                int8;
+typedef short               int16;
+typedef int                 int32;
+typedef long long           int64;
 
 typedef int    bool;
 enum {
